requirements.txt
Pillow==2.6.0
wsgiref==0.1.2
